From e71ad6195ef558f277eb03f1f6560739ce932c6a Mon Sep 17 00:00:00 2001 From: Keno Fischer <keno@juliacomputing.com> Date: Wed, 24 Mar 2021 14:45:48 -0400 Subject: [PATCH] Refactor /proc/meminfo code to use dynamic detection
At the moment, the build system abuses a header check to make a
build-time determination about the existence of /proc/meminfo.
This is in general not idea, because it breaks cross compilation
(e.g. if we're compiling from Linux, which has /proc/meminfo to
Windows, which does not). It's also bad, because the existence
of /proc/meminfo is a transient condition. E.g. a Linux distribution
may sandbox its compile process without acess to /proc, but otherwise
make it available. This removes the compile time check and refactors
the rest of the code to dynamically check for /proc/meminfo. Where
other dynamic fall backs exist, I tried to make sure to maintain them
to avoid breaking users that may not have /proc/meminfo mounted.

While I'm here, also fix two instances of an incorrect case of Windows.h
(most uses in the code base are correct). File systems on windows are
often case insensitive, which is why these kinds of things can easily
slip in (as well as being incorrect in a myriad of documentation and
tutorials) - nevertheless the official headers are all lower case.